+++ src/place/place.module.ts
@@ -1,10 +1,11 @@
 import { Module } from "@nestjs/common";
 
+import { SharedModule } from "../shared/shared.module";
 import { SearchService } from "./kakaoMapSearch/search.service";
 import { PlaceResolver } from "./place.resolver";
 
 @Module({
-  imports: [],
+  imports: [SharedModule],
   providers: [SearchService, PlaceResolver],
   exports: [SearchService],
 })
